# Conflicts:
#   test/data/zentao/processor.php

+++ module/testcase/control.php
@@ -894,20 +894,18 @@ class testcase extends control
             {
                 $branchTagOption[$branchInfo->id] = $branchInfo->name . ($branchInfo->status == 'closed' ? ' (' . $this->lang->branch->statusList['closed'] . ')' : '');
             }
+
             $this->view->productID        = $productID;
             $this->view->product          = $product;
             $this->view->branchTagOption  = $branchTagOption;
             $this->view->productName      = $this->products[$productID];
             $this->view->moduleOptionMenu = $moduleOptionMenu;
-            $this->view->stories          = $this->story->getProductStoryPairs($productID, $case->branch);
+            $this->view->stories          = $this->story->getProductStoryPairs($productID, $case->branch, 0, 'all','id_desc', 0, 'full', 'story', false);
         }
         $forceNotReview = $this->testcase->forceNotReview();
         if($forceNotReview) unset($this->lang->testcase->statusList['wait']);
-        $position[]      = $this->lang->testcase->common;
-        $position[]      = $this->lang->testcase->edit;
 
         $this->view->title           = $title;
-        $this->view->position        = $position;
         $this->view->currentModuleID = $case->module;
         $this->view->users           = $this->user->getPairs('noletter');
         $this->view->case            = $case;
@@ -1868,6 +1866,8 @@ class testcase extends control
         if($_POST)
         {
             $this->testcase->createFromImport($productID, (int)$branch);
+            if(dao::isError()) return $this->send(array('result' => 'fail', 'message' => dao::getError()));
+
             if($this->post->isEndPage)
             {
                 unlink($tmpFile);
